requirements.txt
discord.py
asyncpg
git+https://github.com/Rapptz/discord-ext-menus
